+++ src/spyglass/common/common_dandi.py
@@ -9,7 +9,7 @@ import pynwb
 from fsspec.implementations.cached import CachingFileSystem
 
 from spyglass.common.common_usage import Export, ExportSelection
-from spyglass.settings import export_dir
+from spyglass.settings import export_dir, raw_dir
 from spyglass.utils import SpyglassMixin, logger
 from spyglass.utils.sql_helper_fn import SQLDumpHelper
 
@@ -21,7 +21,8 @@ try:
     from dandi.consts import known_instances
     from dandi.dandiapi import DandiAPIClient
     from dandi.metadata.nwb import get_metadata
-    from dandi.organize import OrganizeInvalid
+    from dandi.organize import CopyMode, FileOperationMode, OrganizeInvalid
+    from dandi.pynwb_utils import nwb_has_external_links
     from dandi.validate_types import Severity
 
 except (ImportError, ModuleNotFoundError) as e:
@@ -31,8 +32,11 @@ except (ImportError, ModuleNotFoundError) as e:
         DandiAPIClient,
         get_metadata,
         OrganizeInvalid,
+        CopyMode,
+        FileOperationMode,
         Severity,
-    ) = [None] * 6
+        nwb_has_external_links,
+    ) = [None] * 9
     logger.warning(e)
 
 
@@ -55,6 +59,12 @@ class DandiPath(SpyglassMixin, dj.Manual):
 
     def has_file_path(self, file_path: str) -> bool:
         return bool(self & self.key_from_path(file_path))
+
+    def raw_from_path(self, file_path) -> dict:
+        return {"filename": Path(file_path).name.replace("_.nwb", ".nwb")}
+
+    def has_raw_path(self, file_path: str) -> bool:
+        return bool(self & self.raw_from_path(file_path))
 
     def fetch_file_from_dandi(
         self, key: dict = None, nwb_file_path: str = None
@@ -99,6 +109,7 @@ class DandiPath(SpyglassMixin, dj.Manual):
         dandiset_id: str,
         dandi_api_key: str = None,
         dandi_instance: str = "dandi",
+        skip_raw_files: bool = False,
     ):
         """Compile a Dandiset from the export.
         Parameters
@@ -112,6 +123,8 @@ class DandiPath(SpyglassMixin, dj.Manual):
             DANDI_API_KEY is set.
         dandi_instance : str, optional
             What instance of Dandi the dandiset is on. Defaults to dev server.
+        skip_raw_files : bool, optional
+            Dev tool to skip raw files in the export. Defaults to False.
         """
         key = (Export & key).fetch1("KEY")
         paper_id = (Export & key).fetch1("paper_id")
@@ -149,9 +162,14 @@ class DandiPath(SpyglassMixin, dj.Manual):
 
         os.makedirs(destination_dir, exist_ok=False)
         for file in source_files:
-            if not os.path.exists(
-                f"{destination_dir}/{os.path.basename(file)}"
-            ):
+            if os.path.exists(f"{destination_dir}/{os.path.basename(file)}"):
+                continue
+            if skip_raw_files and raw_dir in file:
+                continue
+            # copy the file if it has external links so can be safely edited
+            if nwb_has_external_links(file):
+                shutil.copy(file, f"{destination_dir}/{os.path.basename(file)}")
+            else:
                 os.symlink(file, f"{destination_dir}/{os.path.basename(file)}")
 
         # validate the dandiset
@@ -166,11 +184,16 @@ class DandiPath(SpyglassMixin, dj.Manual):
 
         # organize the files in the dandiset directory
         dandi.organize.organize(
-            destination_dir, dandiset_dir, invalid=OrganizeInvalid.WARN
+            destination_dir,
+            dandiset_dir,
+            update_external_file_paths=True,
+            invalid=OrganizeInvalid.FAIL,
+            media_files_mode=CopyMode.SYMLINK,
+            files_mode=FileOperationMode.COPY,
         )
 
         # get the dandi name translations
-        translations = translate_name_to_dandi(destination_dir)
+        translations = lookup_dandi_translation(destination_dir, dandiset_dir)
 
         # upload the dandiset to the dandi server
         if dandi_api_key:
@@ -212,7 +235,7 @@ class DandiPath(SpyglassMixin, dj.Manual):
             docker_id=None,
             spyglass_version=spyglass_version,
         )
-        sql_dump.write_mysqldump(self & key, file_suffix="_dandi")
+        sql_dump.write_mysqldump([self & key], file_suffix="_dandi")
 
 
 def _get_metadata(path):
@@ -241,6 +264,7 @@ def translate_name_to_dandi(folder):
     dict
         dictionary of filename to dandi_path translations
     """
+
     files = Path(folder).glob("*")
     metadata = list(map(_get_metadata, files))
     metadata, skip_invalid = dandi.organize.filter_invalid_metadata_rows(
@@ -253,6 +277,39 @@ def translate_name_to_dandi(folder):
         {"filename": Path(file["path"]).name, "dandi_path": file["dandi_path"]}
         for file in metadata
     ]
+
+
+def lookup_dandi_translation(source_dir: str, dandiset_dir: str):
+    """Get the dandi_path for each nwb file in the source_dir from
+    the organized dandi directory
+
+    Parameters
+    ----------
+    source_dir : str
+        location of the source files
+    dandiset_dir : str
+        location of the organized dandiset directory
+
+    Returns
+    -------
+    dict
+        dictionary of filename to dandi_path translations
+    """
+    # get the obj_id and dandipath for each nwb file in the dandiset
+    dandi_name_dict = {}
+    for dandi_file in Path(dandiset_dir).rglob("*.nwb"):
+        dandi_path = dandi_file.relative_to(dandiset_dir).as_posix()
+        with pynwb.NWBHDF5IO(dandi_file, "r") as io:
+            nwb = io.read()
+            dandi_name_dict[nwb.object_id] = dandi_path
+    # for each file in the source_dir, lookup the dandipath based on the obj_id
+    name_translation = {}
+    for file in Path(source_dir).glob("*"):
+        with pynwb.NWBHDF5IO(file, "r") as io:
+            nwb = io.read()
+            dandi_path = dandi_name_dict[nwb.object_id]
+            name_translation[file.name] = dandi_path
+    return name_translation
 
 
 def validate_dandiset(