+++ src/spyglass/spikesorting/v1/metric_curation.py
@@ -260,10 +260,13 @@ class MetricCuration(SpyglassMixin, dj.Computed):
         os.makedirs(waveforms_dir, exist_ok=True)
 
         logger.info("Extracting waveforms...")
+
+        # Extract non-sparse waveforms by default
+        waveform_params.setdefault("sparse", False)
+
         waveforms = si.extract_waveforms(
             recording=recording,
             sorting=sorting,
-            sparse=waveform_params.get("sparse", False),
             folder=waveforms_dir,
             overwrite=True,
             **waveform_params,